S10P22A402-219[fix] : MenuCategory 오타 수정

+++ piepay-api/src/main/java/com/pay/pie/domain/menu/entity/Menu.java
@@ -44,7 +44,7 @@ public class Menu extends BaseEntity {
 
 	@NotNull
 	@Column(name = "menu_price", nullable = false)
-	private Integer menuPrice;
+	private Long menuPrice;
 
 	@NotNull
 	@Column(name = "menu_category", nullable = false)
